Make query parsing reuse code better across different Django versions
