+++ djangae/db/backends/appengine/parsers/version_19.py
@@ -1,0 +1,5 @@
+from .base import BaseParser
+
+
+class Parser(BaseParser):
+    pass
