+++ djangae/db/backends/appengine/query.py
@@ -5,13 +5,10 @@ import re
 import datetime
 
 from itertools import chain, imap
-from django.core.exceptions import FieldError
-from django.db.models.fields import FieldDoesNotExist
 from django.db.models.sql.datastructures import EmptyResultSet
 
 from django.db.models import AutoField
 
-from django.db.models.query import QuerySet
 
 try:
     from django.db.models.query import FlatValuesListIterable
@@ -27,15 +24,12 @@ except ImportError:
     class ValuesListQuerySet(object):
         pass
 
-from django.db.models.expressions import Star
-
 from django.db import NotSupportedError
 from djangae.db.backends.appengine.indexing import (
     REQUIRES_SPECIAL_INDEXES,
     add_special_index,
 )
 
-from djangae import environment
 from djangae.db.backends.appengine import POLYMODEL_CLASS_ATTRIBUTE
 from djangae.db.utils import (
     get_top_concrete_parent,
@@ -74,17 +68,6 @@ VALID_CONNECTORS = (
 VALID_OPERATORS = (
     '=', '<', '>', '<=', '>=', 'IN'
 )
-
-def _get_concrete_fields_with_model(model):
-    return [
-        (f, f.model if f.model != model else None)
-        for f in model._meta.get_fields()
-        if f.concrete and (
-            not f.is_relation
-            or f.one_to_one
-            or (f.many_to_one and f.related_model)
-        )
-    ]
 
 
 def convert_operator(operator):
@@ -676,521 +659,22 @@ INVALID_ORDERING_FIELD_MESSAGE = (
 )
 
 
-def _extract_ordering_from_query_18(query):
-    from djangae.db.backends.appengine.commands import log_once
-    from django.db.models.expressions import OrderBy, F
+def _get_parser(query, connection=None):
+    version = django.VERSION[:2]
 
-    # Add any orderings
-    if not query.default_ordering:
-        result = list(query.order_by)
+    if version == (1, 8):
+        from djangae.db.backends.appengine.parsers import version_18
+        return version_18.Parser(query, connection)
+    elif version == (1, 9):
+        from djangae.db.backends.appengine.parsers import version_19
+        return version_19.Parser(query, connection)
     else:
-        result = list(query.order_by or query.get_meta().ordering or [])
-
-    if query.extra_order_by:
-        result = list(query.extra_order_by)
-
-        # we need some extra logic to handle dot seperated ordering
-        new_result = []
-        cross_table_ordering = set()
-        for ordering in result:
-            if "." in ordering:
-                dot_based_ordering = ordering.split(".")
-                if dot_based_ordering[0] == query.model._meta.db_table:
-                    ordering = dot_based_ordering[1]
-                elif dot_based_ordering[0].lstrip('-') == query.model._meta.db_table:
-                    ordering = '-{}'.format(dot_based_ordering[1])
-                else:
-                    cross_table_ordering.add(ordering)
-                    continue # we don't want to add this ordering value
-            new_result.append(ordering)
-
-        if len(cross_table_ordering):
-            log_once(
-                DJANGAE_LOG.warning if environment.is_development_environment() else DJANGAE_LOG.debug,
-                "The following orderings were ignored as cross-table orderings are not supported on the datastore: %s", cross_table_ordering
-            )
-
-        result = new_result
-
-    final = []
-
-    opts = query.model._meta
-
-    # Apparently expression ordering is absolute and so shouldn't be flipped
-    # if the standard_ordering is False. This keeps track of which columns
-    # were expressions and so don't need flipping
-    expressions = set()
-
-    for col in result:
-        if isinstance(col, OrderBy):
-            descending = col.descending
-            col = col.expression.name
-            if descending:
-                col = "-" + col
-            expressions.add(col)
-
-        elif isinstance(col, F):
-            col = col.name
-
-        if isinstance(col, (int, long)):
-            # If you do a Dates query, the ordering is set to [1] or [-1]... which is weird
-            # I think it's to select the column number but then there is only 1 column so
-            # unless the ordinal is one-based I have no idea. So basically if it's an integer
-            # subtract 1 from the absolute value and look up in the select for the column (guessing)
-            idx = abs(col) - 1
-            try:
-                field_name = query.select[idx].col.col[-1]
-                field = query.model._meta.get_field(field_name)
-                final.append("-" + field.column if col < 0 else field.column)
-            except IndexError:
-                raise NotSupportedError("Unsupported order_by %s" % col)
-        elif col.lstrip("-") == "pk":
-            pk_col = "__key__"
-            final.append("-" + pk_col if col.startswith("-") else pk_col)
-        elif col == "?":
-            raise NotSupportedError("Random ordering is not supported on the datastore")
-        elif col.lstrip("-").startswith("__") and col.endswith("__"):
-            # Allow stuff like __scatter__
-            final.append(col)
-        elif "__" in col:
-            continue
-        else:
-            try:
-                column = col.lstrip("-")
-
-                # This is really 1.8 only, but I didn't want to duplicate this function
-                # just for this. Suggestions for doing this more cleanly welcome!
-                if column in getattr(query, "annotation_select", {}):
-                    # It's an annotation, if it's a supported one, return the
-                    # original column
-                    annotation = query.annotation_select[column]
-                    name = annotation.__class__.__name__
-
-                    # We only support a few expressions
-                    if name not in ("Trunc", "Col", "Date", "DateTime"):
-                        raise NotSupportedError("Tried to order by unsupported expression")
-                    elif name == "Trunc":
-                        column = annotation.lhs.output_field.column
-                    else:
-                        # Retrieve the original column and use that for ordering
-                        if name == "Col":
-                            column = annotation.output_field.column
-                        else:
-                            column = annotation.col.output_field.column
-
-                field = query.model._meta.get_field(column)
-
-                if field.get_internal_type() in (u"TextField", u"BinaryField"):
-                    raise NotSupportedError(INVALID_ORDERING_FIELD_MESSAGE)
-
-                # If someone orders by 'fk' rather than 'fk_id' this complains as that should take
-                # into account the related model ordering. Note the difference between field.name == column
-                # and field.attname (X_id)
-                if field.related_model and field.name == column and field.related_model._meta.ordering:
-                    raise NotSupportedError("Related ordering is not supported on the datastore")
-
-                column = "__key__" if field.primary_key else field.column
-                final.append("-" + column if col.startswith("-") else column)
-            except FieldDoesNotExist:
-                if col in query.extra_select:
-                    # If the column is in the extra select we transform to the original
-                    # column
-                    try:
-                        field = opts.get_field(query.extra_select[col][0])
-                        column = "__key__" if field.primary_key else field.column
-                        final.append("-" + column if col.startswith("-") else column)
-                        continue
-                    except FieldDoesNotExist:
-                        # Just pass through to the exception below
-                        pass
-
-                available = opts.get_all_field_names()
-                raise FieldError("Cannot resolve keyword %r into field. "
-                    "Choices are: %s" % (col, ", ".join(available))
-                )
-
-    # Reverse if not using standard ordering
-    def swap(col):
-        if col.startswith("-"):
-            return col.lstrip("-")
-        else:
-            return "-{}".format(col)
-
-    if not query.standard_ordering:
-        final = [ x if x in expressions else swap(x) for x in final ]
-
-    if len(final) != len(result):
-        diff = set(result) - set(final)
-        log_once(
-            DJANGAE_LOG.warning if environment.is_development_environment() else DJANGAE_LOG.debug,
-            "The following orderings were ignored as cross-table and random orderings are not supported on the datastore: %s", diff
-        )
-
-    return final
-
-
-def _extract_projected_columns_from_query_18(query):
-    result = []
-
-    if query.select:
-        for x in query.select:
-            column = x.target.column
-            result.append(column)
-        return result
-    else:
-        # If the query uses defer()/only() then we need to process deferred. We have to get all deferred columns
-        # for all (concrete) inherited models and then only include columns if they appear in that list
-        only_load = query.get_loaded_field_names()
-        if only_load:
-            for field, model in _get_concrete_fields_with_model(query.model):
-                model = model or query.model
-                try:
-                    if field.column in only_load[model]:
-                        # Add a field that has been explicitly included
-                        result.append(field.column)
-                except KeyError:
-                    # Model wasn't explicitly listed in the only_load table
-                    # Therefore, we need to load all fields from this model
-                    result.append(field.column)
-            return result
-        else:
-            return []
-
-
-def _walk_django_where(query, trunk_callback, leaf_callback, **kwargs):
-    """
-        Walks through a Django where tree. If a leaf node is encountered
-        the leaf_callback is called, otherwise the trunk_callback is called
-    """
-
-    def walk_node(node, **kwargs):
-        negated = kwargs["negated"]
-
-        if node.negated:
-            negated = not negated
-
-        for child in node.children:
-            new_kwargs = kwargs.copy()
-            new_kwargs["negated"] = negated
-            if not getattr(child, "children", []):
-                leaf_callback(child, **new_kwargs)
-            else:
-                new_parent = trunk_callback(child, **new_kwargs)
-
-                if new_parent:
-                    new_kwargs["new_parent"] = new_parent
-
-                walk_node(child, **new_kwargs)
-
-    kwargs.setdefault("negated", False)
-    walk_node(query.where, **kwargs)
-
-def _django_18_query_walk_leaf(node, negated, new_parent, connection, model, compiler):
-    new_node = WhereNode()
-
-    def convert_rhs_op(node):
-        db_rhs = getattr(node.rhs, '_db', None)
-        if db_rhs is not None and db_rhs != connection.alias:
-            raise ValueError(
-                "Subqueries aren't allowed across different databases. Force "
-                "the inner query to be evaluated using `list(inner_query)`."
-            )
-
-        value = node.get_rhs_op(connection, node.rhs)
-        operator = value.split()[0].lower().strip()
-        if operator == 'between':
-            operator = 'range'
-        return operator
-
-    if not hasattr(node, "lhs"):
-        raise NotSupportedError("Attempted probable subquery, these aren't supported on the datastore")
-
-    # Although we do nothing with this. We need to call it as many lookups
-    # perform validation etc.
-    if not hasattr(node.rhs, "_as_sql"): # Don't call on querysets
-        try:
-            node.process_rhs(compiler, connection)
-        except EmptyResultSet:
-            if node.lookup_name == 'in':
-                node.rhs = []
-            else:
-                raise
-
-    # Leaf
-    if hasattr(node.lhs, 'target'):
-        # from Django 1.9, some node.lhs might not have a target attribute
-        # as they might be wrapping date fields
-        field = node.lhs.target
-        operator = convert_rhs_op(node)
-    else:
-        field = node.lhs.lhs.target
-        operator = convert_rhs_op(node)
-
-        # This deals with things like datefield__month__gt=X which means from this point
-        # on, operator will have two parts in that particular case and will probably need to
-        # be dealt with by a special indexer
-        if node.lookup_name != node.lhs.lookup_name:
-            operator = "{}__{}".format(node.lhs.lookup_name, node.lookup_name)
-
-    if get_top_concrete_parent(field.model) != get_top_concrete_parent(model):
-        raise NotSupportedError("Cross-join where filters are not supported on the datastore")
-
-    # Make sure we don't let people try to filter on a text field, otherwise they just won't
-    # get any results!
-
-    if field.db_type(connection) in ("bytes", "text"):
-        raise NotSupportedError("You can't filter on text or blob fields on the datastore")
-
-    if operator == "isnull" and field.model._meta.parents.values():
-        raise NotSupportedError("isnull lookups on inherited relations aren't supported on the datastore")
-
-    lhs = field.column
-
-    if hasattr(node.rhs, "get_compiler"):
-        # This is a subquery
-        raise NotSupportedError("Attempted to run a subquery on the datastore")
-    elif isinstance(node.rhs, ValuesListQuerySet):
-        # We explicitly handle ValuesListQuerySet because of the
-        # common case of pk__in=Something.objects.values_list("pk", flat=True)
-        # this WILL execute another query, but that is to be expected on a
-        # non-relational datastore.
-
-        rhs = [ x for x in node.rhs ] # Evaluate the queryset
-
-    elif isinstance(node.rhs, QuerySet):
-        # In Django 1.9, ValuesListQuerySet doesn't exist anymore, and instead
-        # values_list returns a QuerySet
-        if node.rhs._iterable_class == FlatValuesListIterable:
-            # if the queryset has FlatValuesListIterable as iterable class
-            # then it's a flat list, and we just need to evaluate the
-            # queryset converting it into a list
-            rhs = [ x for x in node.rhs ]
-        else:
-            # otherwise, we try to get the PK from the queryset
-            rhs = [ x.pk for x in node.rhs ]
-    else:
-        rhs = node.rhs
-
-    was_iter = hasattr(node.rhs, "__iter__")
-    rhs = node.get_db_prep_lookup(rhs, connection)[-1]
-    if rhs and not was_iter and hasattr(rhs, "__iter__"):
-        rhs = rhs[0]
-
-    new_node.set_leaf(
-        lhs,
-        operator,
-        rhs,
-        is_pk_field=field==model._meta.pk,
-        negated=negated,
-        lookup_name=node.lookup_name,
-        namespace=connection.ops.connection.settings_dict.get("NAMESPACE"),
-        target_field=field,
-    )
-
-    # For some reason, this test:
-    # test_update_with_related_manager (get_or_create.tests.UpdateOrCreateTests)
-    # ends up with duplicate nodes in the where tree. I don't know why. But this
-    # weirdly causes the datastore query to return nothing.
-    # so here we don't add duplicate nodes, I can't think of a case where that would
-    # change the query if it's under the same parent.
-    if new_node in new_parent.children:
-        return
-
-    new_parent.children.append(new_node)
-
-def _django_18_query_walk_trunk(node, negated, new_parent, **kwargs):
-    new_node = WhereNode()
-    new_node.connector = node.connector
-    new_node.negated = node.negated
-
-    new_parent.children.append(new_node)
-
-    return new_node
-
-
-def _transform_query_18(connection, kind, query):
-    from django.db.models.sql.where import EmptyWhere
-    if isinstance(query.where, EmptyWhere):
-        # Empty where means return nothing!
-        raise EmptyResultSet()
-
-    ret = Query(query.model, kind)
-    ret.connection = connection
-
-    # Add the root concrete table as the source table
-    root_table = get_top_concrete_parent(query.model)._meta.db_table
-    ret.add_source_table(root_table)
-
-    # Extract the ordering of the query results
-    for order_col in _extract_ordering_from_query_18(query):
-        ret.add_order_by(order_col)
-
-    # Extract any projected columns (values/values_list/only/defer)
-    for projected_col in _extract_projected_columns_from_query_18(query):
-        ret.add_projected_column(projected_col)
-
-    # Add any extra selects
-    for col, select in query.extra_select.items():
-        ret.add_extra_select(col, select[0])
-
-    if query.distinct:
-        # This must happen after extracting projected cols
-        ret.set_distinct(list(query.distinct_fields))
-
-    # Process annotations!
-    if query.annotation_select:
-        for k, v in query.annotation_select.items():
-            ret.add_annotation(k, v)
-
-    # Extract any query offsets and limits
-    ret.low_mark = query.low_mark
-    ret.high_mark = query.high_mark
-
-    output = WhereNode()
-    output.connector = query.where.connector
-
-    _walk_django_where(
-        query,
-        _django_18_query_walk_trunk,
-        _django_18_query_walk_leaf,
-        new_parent=output,
-        connection=connection,
-        negated=query.where.negated,
-        model=query.model,
-        compiler=query.get_compiler(connection.alias)
-    )
-
-    # If there no child nodes, just wipe out the where
-    if not output.children:
-        output = None
-
-    ret.where = output
-
-    return ret
-
-
-def _transform_query_19(connection, kind, query):
-    from django.db.models.sql.where import NothingNode, WhereNode as DjangoWhereNode
-
-    # It could either be a NothingNode, or a WhereNode(AND NothingNode)
-    if (
-            isinstance(query.where, NothingNode) or (
-            isinstance(query.where, DjangoWhereNode) and
-            len(query.where.children) == 1 and
-            isinstance(query.where.children[0], NothingNode)
-            )):
-        # Empty where means return nothing!
-        raise EmptyResultSet()
-
-    ret = Query(query.model, kind)
-    ret.connection = connection
-
-    # Add the root concrete table as the source table
-    root_table = get_top_concrete_parent(query.model)._meta.db_table
-    ret.add_source_table(root_table)
-
-    # Extract the ordering of the query results
-    for order_col in _extract_ordering_from_query_18(query):
-        ret.add_order_by(order_col)
-
-    # Extract any projected columns (values/values_list/only/defer)
-    for projected_col in _extract_projected_columns_from_query_18(query):
-        ret.add_projected_column(projected_col)
-
-    # Add any extra selects
-    for col, select in query.extra_select.items():
-        ret.add_extra_select(col, select[0])
-
-    if query.distinct:
-        # This must happen after extracting projected cols
-        ret.set_distinct(list(query.distinct_fields))
-
-    # Process annotations!
-    if query.annotation_select:
-        for k, v in query.annotation_select.items():
-            ret.add_annotation(k, v)
-
-    # Extract any query offsets and limits
-    ret.low_mark = query.low_mark
-    ret.high_mark = query.high_mark
-
-    output = WhereNode()
-    output.connector = query.where.connector
-
-    _walk_django_where(
-        query,
-        _django_18_query_walk_trunk,
-        _django_18_query_walk_leaf,
-        new_parent=output,
-        connection=connection,
-        negated=query.where.negated,
-        model=query.model,
-        compiler=query.get_compiler(connection.alias)
-    )
-
-    # If there no child nodes, just wipe out the where
-    if not output.children:
-        output = None
-
-    ret.where = output
-
-    return ret
-
-
-def get_factory_for_version(factory, version):
-    try:
-        return factory[version]
-    except KeyError:
-        return factory[max(factory.keys())]
-
-
-_FACTORY = {
-    (1, 8): _transform_query_18,
-    (1, 9): _transform_query_19,
-    (1, 10): _transform_query_19, # Same as 1.9
-}
-
-
-def _determine_query_kind_18(query):
-    if query.annotations:
-        if "__count" in query.annotations:
-            field = query.annotations["__count"].input_field
-            if isinstance(field, Star) or field.value == "*":
-                return "COUNT"
-
-    return "SELECT"
-
-
-def _determine_query_kind_19(query):
-    if query.annotations:
-        if "__count" in query.annotations:
-            field = query.annotations["__count"].source_expressions[0]
-            if isinstance(field, Star) or field.value == "*":
-                return "COUNT"
-
-    return "SELECT"
-
-
-_KIND_FACTORY = {
-    (1, 8): _determine_query_kind_18,
-    (1, 9): _determine_query_kind_19,
-    (1, 10): _determine_query_kind_19  # Same as 1.9
-}
-
+        from djangae.db.backends.appengine.parsers import base
+        return base.BaseParser(query, connection)
 
 def transform_query(connection, query):
-    version = django.VERSION[:2]
-    kind = get_factory_for_version(_KIND_FACTORY, version)(query)
-    return get_factory_for_version(_FACTORY, version)(connection, kind, query)
-
-
-_ORDERING_FACTORY = {
-    (1, 8): _extract_ordering_from_query_18,
-    (1, 9): _extract_ordering_from_query_18,  # Same as 1.8
-    (1, 10): _extract_ordering_from_query_18  # Same as 1.8
-}
+    return _get_parser(query, connection).get_transformed_query()
 
 
 def extract_ordering(query):
-    version = django.VERSION[:2]
-    return get_factory_for_version(_ORDERING_FACTORY, version)(query)
+    return _get_parser(query).get_extracted_ordering()
